requirements.txt
appnope==0.1.3
argcomplete==1.10.3
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.8.2
bibtexparser==2.0.0b2
blinker==1.6.2
chardet==3.0.4
click==8.1.7
comm==0.1.4
compressed-rtf==1.0.6
debugpy==1.8.0
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.3
executing==1.2.0
extract-msg==0.28.7
Flask==2.3.3
IMAPClient==2.1.0
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
nltk==3.8.1
olefile==0.46
packaging==23.1
parso==0.8.3
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.16.1
pylatexenc==3.0a21
PyPDF2==3.0.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.22
pyzmq==25.1.1
regex==2023.8.8
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
stack-data==0.6.2
textract==1.6.5
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
wcwidth==0.2.6
Werkzeug==2.3.7
xlrd==1.2.0
XlsxWriter==3.1.3
zipp==3.16.2
